+++ react/index.cjs
@@ -2,3 +2,4 @@
 
 exports = module.exports = require('../commonjs/react/VirtualScroller.js').default
 exports['default'] = require('../commonjs/react/VirtualScroller.js').default
+exports.useVirtualScroller = require('../commonjs/react/useVirtualScroller.js').default