+++ react/index.js
@@ -1,1 +1,2 @@
 export { default } from '../modules/react/VirtualScroller.js'
+export { default as useVirtualScroller } from '../modules/react/useVirtualScroller.js'