+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 <!-- `virtual-scroller`: in `.updateItems()` handle a case when `items.length` is the same, in which case find different items and if those items are rendered then maybe update them on screen and update their height, if the items are past rendered then maybe just discard all item heights past rendered, if the items are before rendered then maybe ignore and it will jump on scroll up which is kinda acceptable. -->
 
+1.15.1 / 30.11.2025
+===================
+
+* Added `useVirtualScroller()` hook that is exported from `virtual-scroller/react` subpackage.
+* Un-deprecated `tbody: boolean` property because it's used in the new hook.
+
 1.15.0 / 30.11.2025
 ===================